change 8 bit to 16 bit for address

+++ src/LCD_VU.h
@@ -151,7 +151,7 @@ class LCD_VU {
 		void drawBar20(short data, short peakData, short row);
 
 	public:
-		LCD_VU(uint8_t address, uint8_t col, uint8_t row, byte audioPinLeft, byte audioPinRight);
+		LCD_VU(uint16_t address, uint8_t col, uint8_t row, byte audioPinLeft, byte audioPinRight);
 		void init();
 		void loop();
 		void setCursor(uint8_t col, uint8_t row);
